refactor: AI 아티클 사용자 조회 방식 정리
